Added level 3 and 4, added checkpoint mapObject, edited Spikes
Added level 3, based on platforming
Added level 4, based on spikes

Added the checkpoint mapObject, which when touched changes your spawn
position. Also changed spawnX/Y to shorts to allow for larger levels to
have checkpoints

Edited the drawn size if spikes, to make terrible hitbox less apparent

+++ TheBounceArduboy/Player.cpp
@@ -96,7 +96,7 @@ void PlayerClass::update(Arduboy& arduboy)
 	y += yVelocity;
 	x += xVelocity;
 }
-void PlayerClass::respawn(byte spawnX, byte spawnY)
+void PlayerClass::respawn(short spawnX, short spawnY)
 {
 	x = spawnX;
 	y = spawnY;
